add more widgets
