+++ app/javascript/components/widgets/climate/emissions-plantations/selectors.js
@@ -5,8 +5,8 @@ import { formatNumber } from 'utils/format';
 const getData = state => state.data;
 const getLocationName = state => state.locationName;
 const getColors = state => state.colors;
-const getSentences = state => state.config && state.config.sentences;
-const getTitle = state => state.config.title;
+const getSentences = state => state.sentences;
+const getTitle = state => state.title;
 const getSettings = state => state.settings;
 
 export const parseData = createSelector(
