--- app/javascript/components/widgets/climate/emissions-plantations/settings.js
@@ -1,7 +1,0 @@
-export default {
-  forestType: 'ifl',
-  threshold: 30,
-  startYear: 2013,
-  endYear: 2018,
-  unit: 'co2LossByYear'
-};
