--- app/javascript/components/widgets/climate/future-carbon-gains/config.js
@@ -1,57 +1,0 @@
-export default {
-  widget: 'futureCarbonGains',
-  title: {
-    initial: 'Annual tree cover loss by dominant driver in {location}',
-    global: 'Global annual tree cover loss by dominant driver'
-  },
-  categories: ['climate'],
-  colors: 'climate',
-  types: ['country'],
-  admins: ['adm0'],
-  options: {
-    units: ['co2Gain', 'cGain']
-  },
-  metaKey: 'potential_tree_biomass_gain',
-  sortOrder: {
-    summary: 1,
-    forestChange: 1,
-    global: 1
-  },
-  datasets: [
-    // potential carbon gains
-    {
-      dataset: 'b7a34457-1d8a-456e-af46-876e0b42fb96',
-      layers: ['fffa76d3-5008-48b7-afeb-2c7054548f2e']
-    }
-  ],
-  analysis: true,
-  sentences: {
-    initial:
-      'In {location}, potential carbon sequestration may reach {amount} of {variable} by {maxYear}.'
-  },
-  whitelists: {
-    adm0: [
-      'ARG',
-      'BLZ',
-      'BOL',
-      'BRA',
-      'COL',
-      'CRI',
-      'CUB',
-      'ECU',
-      'SLV',
-      'GTM',
-      'GUY',
-      'HND',
-      'JAM',
-      'MEX',
-      'NIC',
-      'PAN',
-      'PRY',
-      'PER',
-      'PRI',
-      'SUR',
-      'VEN'
-    ]
-  }
-};
