--- app/javascript/components/widgets/climate/future-carbon-gains/settings.js
@@ -1,3 +1,0 @@
-export default {
-  unit: 'co2Gain'
-};
